feat: implement authentication and token management decorators, and restructure routes for user and connection management
