+++ src/server/decorators/authenticate.ts
@@ -1,0 +1,9 @@
+import {FastifyPluginAsync} from 'fastify';
+
+const plugin: FastifyPluginAsync = async server => {
+  server.decorate('authenticate', async request => {
+    await request.jwtVerify();
+  });
+};
+
+export default plugin;
