+++ src/server/decorators/authenticateRefresh.ts
@@ -1,0 +1,9 @@
+import {FastifyPluginAsync} from 'fastify';
+
+const plugin: FastifyPluginAsync = async server => {
+  server.decorate('authenticateRefresh', async request => {
+    await request.jwtVerify({onlyCookie: true});
+  });
+};
+
+export default plugin;
