+++ src/server/routes/api/account/connections/delete.ts
@@ -1,19 +1,18 @@
 import {FastifyPluginAsync} from 'fastify';
 import SQL from 'sql-template-strings';
-import fp from 'fastify-plugin';
 
 const plugin: FastifyPluginAsync = async server => {
-  server.post(
-    '/api/auth/logout',
+  server.delete(
+    '',
     {onRequest: server.authenticate},
     async (request, reply) => {
       await server.db.run(
-        SQL`DELETE FROM connections WHERE id = ${request.connection}`,
+        SQL`DELETE FROM connections WHERE id != ${request.connection} AND user_id = ${request.user.id}`,
       );
 
-      return reply.clearCookie('refreshToken').code(204).send();
+      return reply.code(204).send();
     },
   );
 };
 
-export default fp(plugin);
+export default plugin;
