+++ src/server/routes/api/auth/signup/post.ts
@@ -2,7 +2,6 @@ import {FastifyPluginAsyncJsonSchemaToTs} from '@fastify/type-provider-json-sche
 import SQL from 'sql-template-strings';
 import capitalize from '#lib/capitalize';
 import {errorCodes} from 'fastify';
-import fp from 'fastify-plugin';
 import hash from '#lib/hash';
 
 interface ValidationItemParams {
@@ -108,29 +107,25 @@ const plugin: FastifyPluginAsyncJsonSchemaToTs = async server => {
     throw {...error, ...{field: 'username'}};
   }
 
-  server.post(
-    '/api/auth/signup',
-    {schema, attachValidation: true},
-    async (request, reply) => {
-      formatValidationError(request.validationError);
+  server.post('', {schema, attachValidation: true}, async (request, reply) => {
+    formatValidationError(request.validationError);
 
-      const {username} = request.body;
-      await checkUsername(username);
+    const {username} = request.body;
+    await checkUsername(username);
 
-      const password = hash(request.body.password);
+    const password = hash(request.body.password);
 
-      const {lastID: id} = await server.db.run(
-        SQL`INSERT INTO users(username, password) VALUES(${username}, ${password})`,
-      );
-      if (!id) throw new Error('Failed to create user');
+    const {lastID: id} = await server.db.run(
+      SQL`INSERT INTO users(username, password) VALUES(${username}, ${password})`,
+    );
+    if (!id) throw new Error('Failed to create user');
 
-      const {accessToken, refreshToken} = await request.generateTokens(id);
-      return reply
-        .setCookie('refreshToken', refreshToken)
-        .send({accessToken})
-        .code(201);
-    },
-  );
+    const {accessToken, refreshToken} = await request.generateTokens(id);
+    return reply
+      .setCookie('refreshToken', refreshToken)
+      .send({accessToken})
+      .code(201);
+  });
 };
 
-export default fp(plugin);
+export default plugin;
